backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.1.8
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.1.1
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
waitress==3.0.2
Werkzeug==3.1.3
zipp==3.23.0
